chores: extract to package
