+++ wazo_sdk/commands/chores_.py
@@ -2,9 +2,15 @@
 # SPDX-License-Identifier: GPL-3.0+
 
 import os
-import subprocess
 
 from cliff.command import Command
+
+from .chores.chore import Chore
+
+# Those classes need to be imported to be listed
+from .chores.docker import DockerChore
+from .chores.authors import AuthorsChore
+
 
 ARCHIVES = set(
     (
@@ -71,115 +77,9 @@ IGNORED = set(
 )
 
 
-class Chore:
-    name = 'undefined'
-
-    @classmethod
-    def is_applicable(cls, repo_path):
-        return True
-
-    @classmethod
-    def is_dirty(cls, repo_path):
-        return True
-
-    @classmethod
-    def print_dirty_details(cls, repo_path, repo_name):
-        print(repo_name)
-
-
 class NoSuchChore(ValueError):
     def __init__(self, name):
         self.chore_name = name
-
-
-class DockerChore(Chore):
-    name = 'docker'
-
-    @classmethod
-    def is_applicable(cls, repo_path):
-        return has_dockerfile(repo_path)
-
-    @classmethod
-    def is_dirty(cls, repo_path):
-        return needs_split_dockerfile(repo_path)
-
-    @classmethod
-    def print_dirty_details(cls, repo_path, repo_name):
-        print(dockerfile_path(repo_name))
-
-
-def dockerfile_path(repo_path):
-    return os.path.join(repo_path, 'Dockerfile')
-
-
-def has_dockerfile(repo_path):
-    return os.path.isfile(dockerfile_path(repo_path))
-
-
-def needs_split_dockerfile(repo_path):
-    return has_one_dockerfile_from(repo_path) and has_requirements_txt(repo_path)
-
-
-def has_one_dockerfile_from(repo_path):
-    try:
-        command = [
-            'grep',
-            '--ignore-case',
-            '--count',
-            '^FROM',
-            dockerfile_path(repo_path),
-        ]
-        return subprocess.check_output(command).decode('utf-8').strip() == '1'
-    except subprocess.CalledProcessError as e:
-        if e.returncode == 1:
-            return False
-        raise
-
-
-def has_requirements_txt(repo_path):
-    command = [
-        'grep',
-        '--quiet',
-        '--ignore-case',
-        'requirements.txt',
-        dockerfile_path(repo_path),
-    ]
-    return subprocess.run(command).returncode == 0
-
-
-class AuthorsChore(Chore):
-    name = 'authors'
-
-    @classmethod
-    def is_applicable(cls, repo_path):
-        return has_authors_file(repo_path)
-
-    @classmethod
-    def is_dirty(cls, repo_path):
-        return not has_wazo_author(repo_path)
-
-    @classmethod
-    def print_dirty_details(cls, repo_path, repo_name):
-        print(authors_path(repo_name))
-
-
-def authors_path(repo_path):
-    return os.path.join(repo_path, 'AUTHORS')
-
-
-def has_authors_file(repo_path):
-    return os.path.isfile(authors_path(repo_path))
-
-
-def has_wazo_author(repo_path):
-    command = [
-        'grep',
-        '--quiet',
-        '--ignore-case',
-        'Wazo Communication Inc.',
-        authors_path(repo_path),
-    ]
-    return subprocess.run(command).returncode == 0
 
 
 class ChoreList(Command):
